requirements.txt
Django==2.0.7
djangorestframework==3.8.2
geographiclib==1.49
geopy==1.15.0
pytz==2018.5
